// packages/hooks/src/use-action/type.ts
import type { DocumentNode } from "@apollo/client";
import React from "react";

export interface IActionResult {
  name?: string;
  total: number;
  current: number;
  success: number;
  suspended: number;
  running: number;
  fail: number;
  exception: number;
  inventory?: any;
  createDate?: number;
  messageDestroyed?: boolean;
  actionId?: string;
  successMessage?: string | React.ReactNode;
  progress?: number;
}

export interface ITaskResult {
  current: number;
  inventory: any;
  error: any;
  id?: string;
}

interface MiddleStateData {
  type: string;
  uuids: string[];
  field: string;
  data: any;
}

export interface IActionParams<T = any> {
  mutation: DocumentNode;
  name?: string;
  total: number;
  payload: T;
  middleState?: MiddleStateData | MiddleStateData[];
  refetchPolicy?: "progress" | "finish";
  type?: string;
  // 强制执行 onProgress，onFinish 回调。只有在回调中不修改状态时，才可以设置为 true。
  forceRunCallback?: boolean;
  successMessage?: string | React.ReactNode;
  onProgress?: (result: ITaskResult) => void;
  onFinish?: (result: IActionResult) => void;
}

export enum MiddleState {
  Running = "Running",
}